Use XmlUtils instead of Color#getColor
Bug: 156505010
Test: make

Color is moving to a mainline module, so its @hide APIs are no longer
accessible outside its module. (Further, the method is being removed
because it is unused inside the module.) This class do not need the
Color-specific portion of that method (looking up colors like "blue" in
a map); just call XmlUtils directly.

Change-Id: I27f88168c254ac1e53c800695670f9ea7bbf2c74