Add the support for fragment use in carsetupwizard library
* modify setContentFragment & setContentFragmentWithBackstack
in the BaseSetupWizardActivity to support the fragment use
in split-nav implementation. It renders the fragment in the
viewStubContent instead of appending it.
* keep backward compatibility. It's backward compatible because
until splitnav is enabled, the user won't see the CX change. Once
split-nav is enabled, fragment renders within view stub, instead of being
appending into view, which is incorrect rendering.

Test: Unit tests by modifying BaseCompatActivityTest
Test: MANUALLY test for several different cases, including edge cases.

BUG: b/203109230
Change-Id: Iadf19e2c3b26487ded460705e0d5766bea8a87fb