Add API to provide complexity validation error
Currently there is no method for SUW to retrieve the
feedback String when user password/PIN/pattern does
not meet the complexity requirement.

We create a new API checkValidLockForMessage to
retrieve the respective Strings shown to users in
Settings when password does not meet the requirement.

Additionally, we update LIBRARY_VERSION to 2.

Bug: 268506426
Test: manual - build passes when InitialSetupService
in CarSettings implements the method

Change-Id: I37ca7d0b5eec178f993770ba806fc428b7eda0c8
Merged-In: I37ca7d0b5eec178f993770ba806fc428b7eda0c8